Bugfix: learning model was using both server time and client time for calculations
System.getTimeMillis() uses server time and is platform specific. The SS model used server time for start time and requesting a trial, but client time for setting time points.
This causes tiles to decay too fast as they show their activation from the past (in case the server time is behind on the client time).
Fixes:
- The players send their time to the server when they start/join a game, so now the model will use this timestamp as start time.
- The players send their time to the server when they request a trial, so now the model will send the activations back at exactly the time when the user requested it.